Update Fantasy.css (#1169)
* Update Fantasy.css

File Manager Now matches with theme as well as a mobile login changes for theme and fix for Minecraft Module Plugin search box to now show white text and not black text

* Update Fantasy.css

* Update Fantasy.css

Side Menu Hover `#sideMenu a:hover` now is part of the theme. Didn't think about it till now.